Improve group and package names
