+++ extension/integration-tests/src/test/java/io/quarkiverse/observability/minecraft/it/MinecrafterResourceIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.observability.minecraft.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class MinecrafterResourceIT extends MinecrafterResourceTest {
+}
