--- extension/integration-tests/src/test/java/org/acme/minecrafter/it/MinecrafterResourceIT.java
@@ -1,7 +1,0 @@
-package org.acme.minecrafter.it;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class MinecrafterResourceIT extends MinecrafterResourceTest {
-}
