--- extension/runtime/src/main/java/org/acme/minecrafter/runtime/MinecraftLog.java
@@ -1,13 +1,0 @@
-package org.acme.minecrafter.runtime;
-
-import jakarta.interceptor.InterceptorBinding;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@InterceptorBinding
-@Target({ElementType.ANNOTATION_TYPE, ElementType.TYPE, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface MinecraftLog {
-}
