+++ sample/pom.xml
@@ -58,8 +58,8 @@
             <artifactId>quarkus-rest-jackson</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.acme</groupId>
-            <artifactId>minecrafter</artifactId>
+            <groupId>io.quarkiverse.observability.minecraft</groupId>
+            <artifactId>quarkus-observability-minecraft</artifactId>
             <version>1.0.0-SNAPSHOT</version>
         </dependency>
         <dependency>
